RELAY-100 feat: replace hashStore with repository

+++ cmd/utils/network/printers.go
@@ -124,7 +124,7 @@ func printHeaderTable(header entity.ValidatorSetHeader) string {
 	headerTableData := pterm.TableData{
 		{"Field", "Value"},
 		{"Version", strconv.FormatUint(uint64(header.Version), 10)},
-		{"Epoch", strconv.FormatUint(header.Epoch, 10)},
+		{"Epoch", strconv.FormatUint(uint64(header.Epoch), 10)},
 		{"CaptureTimestamp", fmt.Sprintf("%d (%s)",
 			header.CaptureTimestamp,
 			time.Unix(int64(header.CaptureTimestamp), 0).Format("2006-01-02 15:04:05"),
@@ -178,7 +178,7 @@ func printHeaderWithExtraDataToJSON(validatorSetHeader entity.ValidatorSetHeader
 	jsonHeaderData := jsonHeader{
 		Version:            validatorSetHeader.Version,
 		ValidatorsSszMRoot: fmt.Sprintf("0x%064x", validatorSetHeader.ValidatorsSszMRoot),
-		Epoch:              validatorSetHeader.Epoch,
+		Epoch:              uint64(validatorSetHeader.Epoch),
 		RequiredKeyTag:     uint8(validatorSetHeader.RequiredKeyTag),
 		CaptureTimestamp:   validatorSetHeader.CaptureTimestamp,
 		QuorumThreshold:    validatorSetHeader.QuorumThreshold.Int,
